Design of "Vælg barn" page
Several classes for components added. The design in these classes could be changed if needed - but using the same components could provide a more cohesive design overall.

CardContainer: Since a lot of components are nested in cards, using this class ensures that they always have same border radius, etc.

CollapsibleListScrollable: this class builds a collapsible list in a scrollable view, that can be used in a single child, such that not the whole page scrolls.

CollapsibleList: this builds a collapsible list based on a list of elements given. So far there is logic for Children/Pupils, but there is placeholder for Schoolclasses and there can be added logic for Users, etc.

CustomScrollBar: This class creates a design for a custom scrollbar. Can be used if the scrollbar should be the same everywhere.

Custom_app_bar: A component for appbars to avoid hardcoding height etc. for each page.

RightIconButton: uses the already existing CustomButton, but changes the layout such that the icon is to the far right, and the title of the button is to the far left.

TextFieldCard: Uses the already existing text field, but nests it in a card to provide a rounded border, etc.

Feel free to use, add new variables, functionality, etc.

Modification of other existing classes by adding variables, to further adjust layout and/or design.

Added placeholder for single-use code logic on the "Vælg barn" page for a parent.