Refactoring of settings page
Added new classes:
editableSettingsTile as a component for the editable text fields specifically used on the settings page.

SettingsData as a model for the data that can be edited and is shown on the settings page.

Refactoring and small additions to the settings page. Add of component customAppBar that already exists on other brancher, to use it on the page.

Comments will be added.